Add logic to remove events that have been removed remotely.
Created a sync worker with an API for looking up remote entities that
have been removed so that they may be removed locally as well.

Since the remote entities are only saved as new events, a new api
was needed to fetch the event ID's that have been removed and delete
our local copies of them.