Fix scrollbar sync issues.
Updated the event subscriber to manage the scrollbar position on event updates
instead of the activity. Created a factory for the subscriber. Updated the
subscriber to properly get and set the vertical position of the list with
the recyclerview.

The event subscriber temporarily contains an unchecked cast for the layout
manager here in order to make this work for now. This should be refactored,
however, following commits will likely revert this back to a listview
making that change defunct, and will hinder the change back to a
listview.