Fix incorrect image loading.
Recycled views were causing an issue on the guest list with async
image loads. Since the listeners keep a reference to the views, but
the views get recycled, the callbacks would occasionally put the
image in the wrong view.
Fixed this issue by keeping a reference to the URL, which is just
a workaround for now until there is more time to refactor this.