Add Location permission
Requested the location permission in order to allow for directions
to be offered to the user when they are not nearby the convention.
This feature is not currently implemented, but will be soon. So,
the permission has been added in advance so that it does not
block updates for users in the near future.