Update Event Models
Updated to use the new event endpoints and information.
Separated the local event model and the API event model so that changes
to the API are easier to manage without a large refactor like this.
Removed the deleted-events check, since our api this year doesn't
provide that functionality, as well as removed the 'since' endpoint, as
that is not provided either.
There are tests for both new models in this change.
Created a persister for comma separated string types, instead of
having the model keep a csv itself, added tests for that.