Snap-scroll to resume on index listview.
When rotating the device, the view attempts to scroll back to the previous
position, however this was using "smooth scroll" which would animate the list
scrolling. Since there can be hundreds of items in the list, this looked
really weird, so I changed it to set the position instead of scrolling
to it.