Add 18/21+ labels into event index.
For events marked as 18/21+ we show a little badge to indicate this
in the index, as a warning.
Removed the linear layout wrapping text on the panel views so that
the badge can float to the right of the title and forced the title to
fit on one line, so it looks good.