Fixes and improvements to Schedule screen
- Fixed broken UI regarding cut off tabs. Padding doesn't interact well
with PagerTabStrip, so that was removed, and a hardcoded height was
substituted, which seems to work well.
- Updated support libraries by a minor version (sorry!) to fix a bug
  with PagerTabStrip where the tab titles wouldn't be initially
  instantiated
  (https://code.google.com/p/android/issues/detail?id=183127)
- Made a few changes (different instantiation in getVuuew,
setOffscreenLimit, FragmentStatePager) to help ensure that that schedule
pages weren't getting reloaded after swiping between pages, which was
super annoying.
- Added loading indicator to schedule page, and hid empty view by
default to attempt to improve experience, since it could be pretty
confusing to reach the page and see "no events" with no indication it
was loading.