Update Thumbnails on guest index
The thumbnails on the guest index used to be using the full image,
that has now been changed to use the thumbnail. This combined with
updating fresco seems to have resolved the memory issue.
Made the new thumbnails focus crop lower on the image so people's
faces will be more centered like the scrim in the details.