Fix layout for badges when recycled
When the panel views got recycled, the title would end up
being the same size as the previous regaurdless of how much text there
was. This would cause the right-side badges to float in a random
looking position.
Fixed this by right-aligning the badges in the panel view
and also made the title max out at two lines.