Force events to return if event day is in the past.
Previously, ended events would only show if the "show hidden events"
option was selected. However, this causes an empty screen after the con
day is over. So it makes more sense to only hide events that are over on
that day, and show all of them again once that event is in the past.